requirements.txt
openai
telethon
fastapi
uvicorn
python-dotenv
duckduckgo_search
unidecode
tiktoken